Correct the type of the selector parameter in cmd_pause

+++ atsha204a/include/cmd.h
@@ -139,7 +139,7 @@ uint8_t cmd_get_nonce(struct io_interface *ioif, const uint8_t *in, size_t in_si
 uint8_t cmd_get_random(struct io_interface *ioif, uint8_t mode, uint8_t *buf,
 		       size_t size);
 
-uint8_t cmd_pause(struct io_interface *ioif, uint16_t selector);
+uint8_t cmd_pause(struct io_interface *ioif, uint8_t selector);
 
 uint8_t cmd_sha(struct io_interface *ioif, const uint8_t *in, size_t in_size,
 		uint8_t *out, size_t out_size);
